refactor: moves react logic into hooks module
