+++ frontend/src/pages/agreements/review/reviewAgreement.hooks.js
@@ -1,17 +1,70 @@
 import * as React from "react";
-import { BLI_STATUS } from "../../../helpers/budgetLines.helpers";
-import { actionOptions } from "./ReviewAgreement.constants";
+import { useSelector } from "react-redux";
+import {
+    useGetAgreementByIdQuery,
+    useGetServicesComponentsListQuery,
+    useUpdateBudgetLineItemMutation
+} from "../../../api/opsAPI";
+import { BLI_STATUS, groupByServicesComponent } from "../../../helpers/budgetLines.helpers";
+import { useIsAgreementEditable, useIsUserAllowedToEditAgreement } from "../../../hooks/agreement.hooks";
+import useAlert from "../../../hooks/use-alert.hooks";
+import useGetUserFullNameFromId from "../../../hooks/user.hooks";
+import useToggle from "../../../hooks/useToggle";
+import { actionOptions, selectedAction } from "./ReviewAgreement.constants";
+import { anyBudgetLinesByStatus, getSelectedBudgetLines, getTotalBySelectedCans } from "./ReviewAgreement.helpers";
 import suite from "./suite";
 
-export const useReviewAgreement = (agreement, isSuccess) => {
+const useReviewAgreement = (agreementId) => {
     const [action, setAction] = React.useState(""); // for the action accordion
     const [budgetLines, setBudgetLines] = React.useState([]);
     const [pageErrors, setPageErrors] = React.useState({});
     const [isAlertActive, setIsAlertActive] = React.useState(false);
     const [mainToggleSelected, setMainToggleSelected] = React.useState(false);
     const [notes, setNotes] = React.useState("");
-
+    const [afterApproval, setAfterApproval] = useToggle(true);
+    const activeUser = useSelector((state) => state.auth.activeUser);
+    const {
+        isSuccess,
+        data: agreement,
+        error: errorAgreement,
+        isLoading: isLoadingAgreement
+    } = useGetAgreementByIdQuery(agreementId, {
+        refetchOnMountOrArgChange: true
+    });
+    const { setAlert } = useAlert();
+    const [updateBudgetLineItem] = useUpdateBudgetLineItemMutation();
     let res = suite.get();
+
+    const groupedBudgetLinesByServicesComponent = agreement?.budget_line_items
+        ? groupByServicesComponent(agreement.budget_line_items)
+        : [];
+
+    const { data: servicesComponents } = useGetServicesComponentsListQuery(agreement?.id);
+    // NOTE: convert page errors about budget lines object into an array of objects
+    const budgetLinePageErrors = Object.entries(pageErrors).filter((error) => error[0].includes("Budget line item"));
+    const budgetLinePageErrorsExist = budgetLinePageErrors.length > 0;
+    const budgetLineErrors = res.getErrors("budget-line-items");
+    const budgetLineErrorsExist = budgetLineErrors.length > 0;
+    const areThereBudgetLineErrors = budgetLinePageErrorsExist || budgetLineErrorsExist;
+    const anyBudgetLinesDraft = anyBudgetLinesByStatus(agreement ?? {}, "DRAFT");
+    const anyBudgetLinePlanned = anyBudgetLinesByStatus(agreement ?? {}, "PLANNED");
+    const changeInCans = getTotalBySelectedCans(budgetLines);
+    const actionOptionsToChangeRequests = {
+        [actionOptions.CHANGE_DRAFT_TO_PLANNED]: selectedAction.DRAFT_TO_PLANNED,
+        [actionOptions.CHANGE_PLANNED_TO_EXECUTING]: selectedAction.PLANNED_TO_EXECUTING
+    };
+    let changeRequestAction = actionOptionsToChangeRequests[action];
+    const isAnythingSelected = getSelectedBudgetLines(budgetLines).length > 0;
+    const isDRAFTSubmissionReady =
+        anyBudgetLinesDraft && action === actionOptions.CHANGE_DRAFT_TO_PLANNED && isAnythingSelected;
+    const isPLANNEDSubmissionReady =
+        anyBudgetLinePlanned && action === actionOptions.CHANGE_PLANNED_TO_EXECUTING && isAnythingSelected;
+    const isSubmissionReady = isDRAFTSubmissionReady || isPLANNEDSubmissionReady;
+
+    const isAgreementStateEditable = useIsAgreementEditable(agreement?.id);
+    const canUserEditAgreement = useIsUserAllowedToEditAgreement(agreement?.id);
+    const isAgreementEditable = isAgreementStateEditable && canUserEditAgreement;
+    const projectOfficerName = useGetUserFullNameFromId(agreement?.project_officer_id);
 
     React.useEffect(() => {
         const newBudgetLines =
@@ -44,6 +97,72 @@ export const useReviewAgreement = (agreement, isSuccess) => {
             setIsAlertActive(false);
         };
     }, [res, isSuccess]);
+
+    const handleSendToApproval = () => {
+        if (anyBudgetLinesDraft || anyBudgetLinePlanned) {
+            const selectedBudgetLines = getSelectedBudgetLines(budgetLines);
+            let selectedBLIsWithStatusAndNotes = [];
+
+            switch (action) {
+                case actionOptions.CHANGE_DRAFT_TO_PLANNED:
+                    selectedBLIsWithStatusAndNotes = selectedBudgetLines.map((bli) => {
+                        return { id: bli.id, status: BLI_STATUS.PLANNED, requestor_notes: notes };
+                    });
+                    break;
+                case actionOptions.CHANGE_PLANNED_TO_EXECUTING:
+                    selectedBLIsWithStatusAndNotes = selectedBudgetLines.map((bli) => {
+                        return {
+                            id: bli.id,
+                            status: BLI_STATUS.EXECUTING,
+                            requestor_notes: notes
+                        };
+                    });
+                    break;
+                default:
+                    break;
+            }
+
+            const currentUserId = activeUser?.id;
+
+            console.log("BLI Package Data:", selectedBudgetLines, currentUserId, notes);
+            console.log("THE ACTION IS:", action);
+
+            let promises = selectedBLIsWithStatusAndNotes.map((budgetLine) => {
+                const { id, data: cleanExistingBLI } = cleanBudgetLineItemForApi(budgetLine);
+                return updateBudgetLineItem({ id, data: cleanExistingBLI })
+                    .unwrap()
+                    .then((fulfilled) => {
+                        console.log("Updated BLI:", fulfilled);
+                    })
+                    .catch((rejected) => {
+                        console.error("Error Updating Budget Line");
+                        console.error({ rejected });
+                        throw new Error("Error Updating Budget Line");
+                    });
+            });
+            Promise.allSettled(promises).then((results) => {
+                let rejected = results.filter((result) => result.status === "rejected");
+                if (rejected.length > 0) {
+                    console.error(rejected[0].reason);
+                    setAlert({
+                        type: "error",
+                        heading: "Error Sending Agreement Edits",
+                        message: "There was an error sending your edits for approval. Please try again.",
+                        redirectUrl: "/error"
+                    });
+                } else {
+                    setAlert({
+                        type: "success",
+                        heading: "Changes Sent to Approval",
+                        // TODO: add Change Requests to alert message
+                        message:
+                            "Your changes have been successfully sent to your Division Director to review. Once approved, they will update on the agreement.",
+                        redirectUrl: "/agreements"
+                    });
+                }
+            });
+        }
+    };
 
     const handleSelectBLI = (bliId) => {
         const newBudgetLines = budgetLines.map((bli) => {
@@ -93,6 +212,29 @@ export const useReviewAgreement = (agreement, isSuccess) => {
         setBudgetLines(newBudgetLines);
     };
 
+    const cleanBudgetLineItemForApi = (data) => {
+        const cleanData = { ...data };
+        if (data.services_component_id === 0) {
+            cleanData.services_component_id = null;
+        }
+        if (cleanData.date_needed === "--") {
+            cleanData.date_needed = null;
+        }
+        const budgetLineId = cleanData.id;
+        delete cleanData.created_by;
+        delete cleanData.created_on;
+        delete cleanData.updated_on;
+        delete cleanData.can;
+        delete cleanData.id;
+        delete cleanData.canDisplayName;
+        delete cleanData.versions;
+        delete cleanData.clin;
+        delete cleanData.agreement;
+        delete cleanData.financialSnapshotChanged;
+
+        return { id: budgetLineId, data: cleanData };
+    };
+
     return {
         action,
         setAction,
@@ -107,7 +249,27 @@ export const useReviewAgreement = (agreement, isSuccess) => {
         mainToggleSelected,
         setMainToggleSelected,
         notes,
-        setNotes
+        setNotes,
+        servicesComponents,
+        groupedBudgetLinesByServicesComponent,
+        handleSendToApproval,
+        areThereBudgetLineErrors,
+        isSubmissionReady,
+        changeRequestAction,
+        changeInCans,
+        anyBudgetLinesDraft,
+        anyBudgetLinePlanned,
+        budgetLineErrorsExist,
+        budgetLineErrors,
+        budgetLinePageErrorsExist,
+        budgetLinePageErrors,
+        errorAgreement,
+        isLoadingAgreement,
+        isAgreementEditable,
+        projectOfficerName,
+        afterApproval,
+        setAfterApproval,
+        agreement
     };
 };
